requirements.txt
asn1crypto==0.24.0
Django==2.2
django-cors-headers==2.5.2
djangorestframework==3.9.2
httplib2==0.9.2
mysqlclient==1.4.1
Pillow==5.1.0
pycrypto==2.6.1
requests==2.18.4
sqlparse==0.3.0
urllib3==1.22
pkg-resources==0.0.0
PyMySQL==0.9.3
